requirements.txt
flask==3.1.1
redis==6.2.0
hiredis==3.2.1
pandas==2.3.1
tabulate==0.9.0
sentence-transformers==5.1.0
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
redisvl==0.8.0
python-dotenv==1.0.0